Improve message when trying to export invalid ids
Also move ImportError due to namespace conflicts with snafu macro.
