+++ rslib/src/import_export/package/colpkg/import.rs
@@ -14,10 +14,10 @@ use zstd::stream::copy_decode;
 use crate::collection::CollectionBuilder;
 use crate::error::FileIoSnafu;
 use crate::error::FileOp;
-use crate::error::ImportError;
 use crate::import_export::package::media::extract_media_entries;
 use crate::import_export::package::media::SafeMediaEntry;
 use crate::import_export::package::Meta;
+use crate::import_export::ImportError;
 use crate::import_export::ImportProgress;
 use crate::import_export::IncrementableProgress;
 use crate::io::atomic_rename;
