+++ rslib/src/import_export/package/media.rs
@@ -25,9 +25,9 @@ use super::MediaEntry;
 use super::Meta;
 use crate::error::FileIoError;
 use crate::error::FileOp;
-use crate::error::ImportError;
 use crate::error::InvalidInputError;
 use crate::import_export::package::colpkg::export::MaybeEncodedWriter;
+use crate::import_export::ImportError;
 use crate::io::atomic_rename;
 use crate::io::filename_is_safe;
 use crate::io::new_tempfile_in;
